refactor RebindManager to have a RebindIteration
introduce new class and new methods for each phase. (no real changes.)
